tambah register

+++ app/Http/Controllers/Admin/AdminUserController.php
@@ -68,7 +68,7 @@ class AdminUserController extends Controller
 
             $user = User::where('uuid', $uuid)->first();
             if (Request()->hasFile('image_profile')) {
-                if (Storage::exists($user->image_profile)) {
+                if ($user->image_profile && Storage::exists($user->image_profile)) {
                     Storage::delete($user->image_profile);
                 }
                 $file_name = $request->image_profile->getClientOriginalName();
